Face Detection application designed in Processing
This is a face detection application designed in Processing and uses
Rekognition api. If you want to execute this program you need to get the
API key and API secret by registering at
http://rekognition.com/register. Once you've got your API key and API
secret, copy and paste it into a .txt file and name it "key.txt" and add
it to the data folder where your program is saved.